LayoutTests/platform/qt-5.0-wk2/http/tests/websocket/tests/hybi/bad-sub-protocol-non-ascii-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket '\u007F'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u0080'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u3042'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uFFFF'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uFEFF'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uFFFE'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uD840\uDC0B'
Test WebSocket bad sub-protocol names by non-ASCII chars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "あ") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￿") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "﻿") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￾") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "𠀋") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS successfullyParsed is true

TEST COMPLETE
